requirements.txt
requests
beautifulsoup4
lxml
mysql-connector-python